rename data file state tests
